refactor: move certs functions to new module
